Rename cs1101s cheatsheet as cs1101s.tex
There is a 20% test for UQF2101I in a few days time. I plan to make a
cheatsheet for that and upload it to github, as I had done for CS1101S.
In that case, it would make more sense to keep both (and future)
cheatsheets in a single 'Cheatsheet' repository, rather than having one
repository per cheatsheet.